+++ src/main/kotlin/blackjack/view/OutputView.kt
@@ -1,17 +1,17 @@
 package blackjack.view
 
-import blackjack.model.BlackJack
-import blackjack.model.Bust
+import blackjack.model.state.BlackJack
+import blackjack.model.state.Bust
 import blackjack.model.Card
 import blackjack.model.Dealer
 import blackjack.model.GameResult
 import blackjack.model.Hand
-import blackjack.model.Hit
+import blackjack.model.state.Hit
 import blackjack.model.Player
 import blackjack.model.PlayerGroup
-import blackjack.model.Running
-import blackjack.model.Running.Companion.INITIAL_CARD_COUNTS
-import blackjack.model.Stay
+import blackjack.model.state.Running
+import blackjack.model.state.Running.Companion.INITIAL_CARD_COUNTS
+import blackjack.model.state.Stay
 
 object OutputView {
     fun printGameSetting(
